bump to v28.0.0-rc.0

+++ include/autd3/driver/firmware/fpga/loop_behavior.hpp
@@ -6,17 +6,14 @@
 namespace autd3::driver {
 
 struct LoopBehavior final {
-  AUTD3_API static inline const native_methods::LoopBehavior Infinite =
-      native_methods::AUTDLoopBehaviorInfinite();
+  AUTD3_API static inline const native_methods::LoopBehavior Infinite = native_methods::AUTDLoopBehaviorInfinite();
 
-  AUTD3_API [[nodiscard]] static native_methods::LoopBehavior Finite(
-      const uint32_t v) {
+  AUTD3_API [[nodiscard]] static native_methods::LoopBehavior Finite(const uint16_t v) {
     if (v == 0) throw AUTDException("Value must be greater than 0");
     return native_methods::AUTDLoopBehaviorFinite(v);
   }
 
-  AUTD3_API static inline const native_methods::LoopBehavior Once =
-      native_methods::AUTDLoopBehaviorOnce();
+  AUTD3_API static inline const native_methods::LoopBehavior Once = native_methods::AUTDLoopBehaviorOnce();
 };
 
 }  // namespace autd3::driver
